+++ source/renderer/app/Routes.js
@@ -9,6 +9,7 @@ import InitialSettingsPage from './containers/profile/InitialSettingsPage';
 import Settings from './containers/settings/Settings';
 import GeneralSettingsPage from './containers/settings/categories/GeneralSettingsPage';
 import WalletsSettingsPage from './containers/settings/categories/WalletsSettingsPage';
+import StakePoolsSettingsPage from './containers/settings/categories/StakePoolsSettingsPage';
 import SupportSettingsPage from './containers/settings/categories/SupportSettingsPage';
 import TermsOfUseSettingsPage from './containers/settings/categories/TermsOfUseSettingsPage';
 import TermsOfUsePage from './containers/profile/TermsOfUsePage';
@@ -92,6 +93,10 @@ export const Routes = withRouter(() => (
             <Route
               path={ROUTES.SETTINGS.WALLETS}
               component={WalletsSettingsPage}
+            />
+            <Route
+              path={ROUTES.SETTINGS.STAKE_POOLS}
+              component={StakePoolsSettingsPage}
             />
             <Route
               path={ROUTES.SETTINGS.TERMS_OF_USE}